+++ apps/ioc-maker-cli/examples/name-collision-example/demo.ts
@@ -1,0 +1,2 @@
+import { container } from "./container.gen";
+const userController = container.resolve("UserController")